fix: comment id model update to long instead of int
* #2927: comment id model update to long instead of int

* #2927: code review fixes (1)

* #2927: code review fixes (2)

* #2927: comment id model update to long instead of int: unit tests fix

* #2927: code review fixes

* Fixed most names of parameters

---------

+++ Octokit.Tests.Integration/Clients/PullRequestReviewCommentsClientTests.cs
@@ -928,7 +928,7 @@ public class PullRequestReviewCommentsClientTests : IDisposable
         int numberToCreate = 2;
         using (var context = await _github.CreateRepositoryContextWithAutoInit(Helper.MakeNameWithTimestamp("PullRequestReviewCommentsReactionTests")))
         {
-            var commentIds = new List<int>();
+            var commentIds = new List<long>();
 
             // Create a test pull request
             var pullRequest = await CreatePullRequest(context);
@@ -967,7 +967,7 @@ public class PullRequestReviewCommentsClientTests : IDisposable
         int numberToCreate = 2;
         using (var context = await _github.CreateRepositoryContextWithAutoInit(Helper.MakeNameWithTimestamp("PullRequestReviewCommentsReactionTests")))
         {
-            var commentIds = new List<int>();
+            var commentIds = new List<long>();
 
             // Create multiple test pull requests
             for (int count = 1; count <= numberToCreate; count++)
@@ -1000,7 +1000,7 @@ public class PullRequestReviewCommentsClientTests : IDisposable
         }
     }
 
-    async Task<int> HelperCreatePullRequestReviewCommentWithReactions(string owner, string repo, PullRequestData pullRequest)
+    async Task<long> HelperCreatePullRequestReviewCommentWithReactions(string owner, string repo, PullRequestData pullRequest)
     {
         var github = Helper.GetAuthenticatedClient();
